+++ src/main/java/org/primal/Primal_Client.java
@@ -21,6 +21,7 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.client.event.*;
+import net.minecraftforge.fml.loading.FMLLoader;
 import org.jetbrains.annotations.NotNull;
 import org.primal.client.layer.HelmetDecorationLayer;
 import org.primal.client.model.helmet_decoration.DeerAntlersModel;
@@ -29,6 +30,7 @@ import org.primal.client.renderer.EntityOnNeckLayer;
 import org.primal.client.renderer.block_entity.*;
 import org.primal.client.renderer.entity.*;
 import org.primal.client.renderer.replaced.*;
+import org.primal.compat.PetCemeteryCompat;
 import org.primal.item.ConchShellItem;
 import org.primal.particle.DreamcatcherAshParticle;
 import org.primal.particle.SnakeSkinFlakeParticle;
@@ -60,8 +62,10 @@ public class Primal_Client {
         if (Primal_Main.COMMON_CONFIG.rabbitModelChange.get())
             EntityRenderers.register(EntityType.RABBIT, RabbitRenderer::new);
 
-        if (Primal_Main.COMMON_CONFIG.wolfModelChange.get())
+        if (Primal_Main.COMMON_CONFIG.wolfModelChange.get()){
             EntityRenderers.register(EntityType.WOLF, WolfRenderer::new);
+            if(FMLLoader.getLoadingModList().getModFileById("pet_cemetery")!=null) PetCemeteryCompat.registerMobsRenderer();
+        }
 
         if (Primal_Main.COMMON_CONFIG.dolphinModelChange.get())
             EntityRenderers.register(EntityType.DOLPHIN, DolphinRenderer::new);