+++ src/main/java/org/primal/client/renderer/replaced/WolfRenderer.java
@@ -2,6 +2,7 @@ package org.primal.client.renderer.replaced;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.animal.Wolf;
 import net.minecraft.nbt.CompoundTag;
@@ -45,6 +46,14 @@ public class WolfRenderer extends GeoReplacedEntityRenderer<Wolf, WolfReplaced> 
         CompoundTag mainTag = new CompoundTag();
         wolf.saveWithoutId(mainTag);
         String variantTag= mainTag.getString("variant");
+
+        if(BuiltInRegistries.ENTITY_TYPE.getKey(this.currentEntity.getType()).getNamespace().equals("pet_cemetery")) {
+            if (BuiltInRegistries.ENTITY_TYPE.getKey(this.currentEntity.getType()).getPath().equals("skeleton_wolf"))
+                return ResourceLocation.fromNamespaceAndPath(Primal_Main.MOD_ID, "textures/entity/wolf/pet_cemetery/skeleton.png");
+
+            if (BuiltInRegistries.ENTITY_TYPE.getKey(this.currentEntity.getType()).getPath().equals("zombie_wolf"))
+                return ResourceLocation.fromNamespaceAndPath(Primal_Main.MOD_ID, "textures/entity/wolf/pet_cemetery/zombie.png");
+        }
 
         return convertWolfVariantToName(variantTag);
     }