+++ src/main/java/org/primal/util/mob_types/CustomFieldGuideState.java
@@ -1,0 +1,8 @@
+package org.primal.util.mob_types;
+
+public interface CustomFieldGuideState {
+
+    boolean hasFieldGuideState();
+
+    void setFieldGuideState(boolean state);
+}